refactor: refactoring RNN models to new standard
